+++ Libraries/LibC/dirent.cpp
@@ -103,7 +103,7 @@ static int allocate_dirp_buffer(DIR* dirp)
         errno = old_errno;
         return new_errno;
     }
-    size_t size_to_allocate = max(st.st_size, 4096);
+    size_t size_to_allocate = max(st.st_size, static_cast<off_t>(4096));
     dirp->buffer = (char*)malloc(size_to_allocate);
     ssize_t nread = syscall(SC_get_dir_entries, dirp->fd, dirp->buffer, size_to_allocate);
     if (nread < 0) {
